GAP-1975: Update 'Sign up for grant updates' Newsletter to use One Login (#65)
* GAP-1975|add support for sub when making changes to newsletter subscription

* GAP-1975|prefer sub instead of email for newsletter

* GAP-1975|fix tests

* GAP-1975|remove duplicate function

* GAP-1975|fix linting

* GAP-1975|fix linting again

---------

+++ src/newsletter/newsletter.controller.ts
@@ -30,13 +30,13 @@ export class NewsletterController {
         return this.newsletterService.findOneById(id);
     }
 
-    @Get('/users/:plainTextEmailAddress/types/:newsletterType')
-    async findOneByEmailAndType(
-        @Param('plainTextEmailAddress') plainTextEmailAddress: string,
+    @Get('/users/:id/types/:newsletterType')
+    async findOneByUserAndType(
+        @Param('id') id: string,
         @Param('newsletterType') type: NewsletterType,
     ) {
-        return this.newsletterService.findOneByEmailAddressAndType(
-            plainTextEmailAddress,
+        return this.newsletterService.findOneBySubOrEmailAddressAndType(
+            id,
             type,
         );
     }
@@ -44,9 +44,10 @@ export class NewsletterController {
     @Post()
     async create(
         @Body('email') plainTextEmailAddress: string,
+        @Body('sub') sub: string,
         @Body('newsletterType') type: NewsletterType,
     ) {
-        return this.newsletterService.create(plainTextEmailAddress, type);
+        return this.newsletterService.create(plainTextEmailAddress, type, sub);
     }
 
     @Delete(':newsletterId')
@@ -57,17 +58,20 @@ export class NewsletterController {
         response.send();
     }
 
-    @Delete('/users/:plainTextEmailAddress/types/:newsletterType')
+    @Delete('/users/:id/types/:newsletterType')
     async deleteByUserAndType(
-        @Param('plainTextEmailAddress') plainTextEmailAddress: string,
+        @Param('id') id: string,
         @Param('newsletterType') type: NewsletterType,
         @Res() response: Response,
         @Query() query: { unsubscribeReference?: string },
     ) {
-        const result = await this.newsletterService.deleteByEmailAddressAndType(
-            plainTextEmailAddress,
-            type,
-        );
+        let result = await this.newsletterService.deleteBySubAndType(id, type);
+        if (result.affected === 0) {
+            result = await this.newsletterService.deleteByEmailAddressAndType(
+                id,
+                type,
+            );
+        }
         if (query?.unsubscribeReference) {
             await this.unsubscribeService
                 .deleteOneById(query.unsubscribeReference)
@@ -81,7 +85,7 @@ export class NewsletterController {
                 });
         }
 
-        result == 0 ? response.status(404) : response.status(204);
+        result.affected === 0 ? response.status(404) : response.status(204);
 
         response.send();
     }
